Move Gallery to components/shared
