+++ Week-1/src/components/shared/Gallery.js
