+++ Week-1/src/components/views/Product/Product.js
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react'
-import Gallery from '~/src/components/Gallery.js'
+import Gallery from '~/src/components/shared/Gallery.js'
 import Portal from '~/src/components/shared/Portal.js'
 import { imageAttributes } from '~/src/shared/helper.js'
 import ProductCard from '~/src/components/Product/Card.js'
